added multi-ticket chomping to throttler, fiddled w batchEnable quotas w Google
