+++ src/main/scala/org/broadinstitute/dsde/workbench/gpalloc/util/Sequentially.scala
@@ -1,0 +1,20 @@
+package org.broadinstitute.dsde.workbench.gpalloc.util
+
+import scala.concurrent.Future
+
+trait Sequentially {
+  //stolen: https://gist.github.com/ryanlecompte/6313683
+  def sequentially[A,T](items: Seq[A])(f: A => Future[T]): Future[Unit] = {
+    items.headOption match {
+      case Some(nextItem) =>
+        val fut = f(nextItem)
+        fut.flatMap { _ =>
+          // successful, let's move on to the next!
+          sequentially(items.tail)(f)
+        }
+      case None =>
+        // nothing left to process
+        Future.successful(())
+    }
+  }
+}
