+++ src/test/resources/reference.conf
@@ -24,11 +24,12 @@ gpalloc {
   opsThrottle = 15
   opsThrottlePerDuration = 10 second
 
-  #There's also a second, undocumented serviceUsage quota of
-  #20 mutate requests per 100 seconds, which applies to calls to
-  #batchEnable specifically.
-  enableThrottle = 1
-  enableThrottlePerDuration = 7 seconds #14.2 rq per 100s
+  #There's also a second serviceUsage quota of 20 mutate tickets
+  #per 100 seconds. A batchEnable call eats 5 mutate tickets.
+  #We got the mutate limit bumped to 500 calls per 100s, i.e.
+  #100 batchEnable calls per 100 seconds, i.e. 1 per second.
+  enableThrottle = 10
+  enableThrottlePerDuration = 15 seconds
 }
 
 mysql {
